Fixes and more tests for the data acces plugin
